Add map support to EncodeAsBlock() and EncodeIntoBody()
This change is motivated by the Nomad Terraform provider which is now
using EncodeAsBlock() to convert from an *api.Job to the Terraform type
system. Since we are already using DecodeBody() to make the conversion
one way it makes sense to use EncodeAsBlock() to make the same
conversion the other way around.

Since it looks like those two functions did not have tests, I added one
for each to test my changes.